// _examples/node-ts/server-fastify/server.gen.ts
export interface ExampleServer<Context = unknown> {
  ping(ctx: Context, req: PingRequest): Promise<PingResponse>
  getUser(ctx: Context, req: GetUserRequest): Promise<GetUserResponse>
  getArticle(ctx: Context, req: GetArticleRequest): Promise<GetArticleResponse>
  uploadAvatar(ctx: Context, req: UploadAvatarRequest): Promise<UploadAvatarResponse>
  downloadAvatar(ctx: Context, req: DownloadAvatarRequest): Promise<Blob | File>
}

//
// Schema types
//

export enum Kind {
  USER = 'USER',
  ADMIN = 'ADMIN'
}

export interface User {
  id: number
  USERNAME: string
  role: Kind
  meta: {[key: string]: any}
  balance: bigint
  createdAt?: string
  extra: Extra
}

export interface Extra {
  info: string
  amount: bigint
  points: Array<bigint>
}

export interface Page {
  num: number
}

export interface GetArticleRequest {
  articleId: number
  byBN: bigint
}

export interface GetArticleResponse {
  title: string
  content?: string
  largeNum: bigint
}

export interface PingRequest {
  counter?: bigint
}

export interface PingResponse {
}

export interface GetUserRequest {
  userId: number
}

export interface GetUserResponse {
  code: number
  user: User
}


export interface UploadAvatarRequest {
  userId: number
  avatar: Blob | File
}

export interface UploadAvatarResponse {
  size: number
  name: string
}

export interface DownloadAvatarRequest {
  userId: number
}
// ...
type WebrpcErrorParams = { name?: string, code?: number, message?: string, status?: number, cause?: string }

export class WebrpcError extends Error {
  code: number
  status: number

  constructor(error: WebrpcErrorParams = {}) {
    super(error.message)
    this.name = error.name || 'WebrpcEndpointError'
    this.code = typeof error.code === 'number' ? error.code : 0
    this.message = error.message || `endpoint error`
    this.status = typeof error.status === 'number' ? error.status : 400
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcError.prototype)
  }

  static new(payload: any): WebrpcError {
    return new this({ message: payload.message, code: payload.code, status: payload.status, cause: payload.cause })
  }
}


export class WebrpcEndpointError extends WebrpcError {
  constructor(error: WebrpcErrorParams = {}) {
    super(error)
    this.name = error.name || 'WebrpcEndpoint'
    this.code = typeof error.code === 'number' ? error.code : 0
    this.message = error.message || `endpoint error`
    this.status = typeof error.status === 'number' ? error.status : 400
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcEndpointError.prototype)
  }
}
// ...
export class WebrpcBadRouteError extends WebrpcError {
  constructor(error: WebrpcErrorParams = {}) {
    super(error)
    this.name = error.name || 'WebrpcBadRoute'
    this.code = typeof error.code === 'number' ? error.code : -2
    this.message = error.message || `bad route`
    this.status = typeof error.status === 'number' ? error.status : 404
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcBadRouteError.prototype)
  }
}